requirements.txt
Django==1.8.7
django-rest-swagger==0.3.4
djangorestframework==3.3.1
Pillow==3.0.0
psycopg2==2.6.1
PyYAML==3.11
wheel==0.24.0
